Neural Network v0.0.1 NeuralNetwork.Connection

Neurons communciate via connections. Connection weights determine the network output and are updated while training occurs. Network capability is represented in the network matrix of weight values.

Functions

connection_for(source_pid, target_pid)

Convenience method that takes two neuron pids, creates a connection, and then returns connection pid.

get(pid)

Return connection by PID

start_link(connection_fields \\ %{})

update(pid, fields)