Neural Network v0.1.0 NeuralNetwork.Layer

List of neurons. The are used to apply behaviors on sets of neurons. A network is made up layers (which are made up of neurons).

Functions

activate(layer_pid, values \\ nil)

Activate all neurons in the layer with a list of values.

add_neurons(layer_pid, neurons)

Add neurons to the layer.

clear_neurons(layer_pid)

Empty out the layer of all neurons.

connect(input_layer_pid, output_layer_pid)

Connect every neuron in the input layer to every neuron in the target layer.

get(pid)

Return a layer by pid.

set_neurons(layer_pid, neurons)

Clear and set neurons in the layer with a given list of neurons.

start_link(layer_fields \\ %{})

train(layer, target_outputs \\ [])

Update all deltas for each neuron.

update(pid, fields)

Update a layer by passing in a pid and a map of fields to update.