Neural Network v0.1.0 NeuralNetwork.Network

Contains layers which makes up a matrix of neurons.

Functions

activate(network, input_values)

Activate the network given list of input values.

get(pid)

Return the network by pid.

start_link(layer_sizes \\ [])

Pass in layer sizes which will generate the layers for the network. The first number represents the number of neurons in the input layer. The last number represents the number of neurons in the output layer. [Optionally] The middle numbers represent the number of neurons for hidden layers.

train(network, target_outputs)

Set the network error and output layer’s deltas propagate them backward through the network.

The input layer is skipped (no use for deltas).

update(pid, fields)

Update the network layers.